requirements.txt
pydantic-settings
opentelemetry-exporter-otlp-proto-http
